requirements.txt
# Core deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Configuration and utilities
PyYAML>=6.0
tqdm>=4.65.0

# Image processing
matplotlib>=3.7.0
Pillow>=9.5.0

# Tokenization (using TensorFlow's Keras preprocessing)
tensorflow>=2.12.0

